Fix maven-scr-plugin does not generate XML for DS annotations in test bundles (#395)
The maven-resources-plugin is used to copy back the generated XML to the OSGI-INF directory so it is on the classpath in tests.
The outputDirectory configuration parameter of the maven-scr-plugin is not used because when it is set to ${project.basedir} it empties the OSGI-INF dir and would remove non-generated SCR XML files.

Fixes #318

Signed-off-by: Wouter Born <eclipse@maindrain.net>
GitOrigin-RevId: 28c5922149921f41c7b1a99e7545828e9eb01660