contains + get can be replaced by get and a null check (#1269)
contains + get will do its job as long as all collection access uses the
same synchronization locks and as long as every put ensures there will
never by a null value be inserted.

If the map implementation checks if a key is available, we can also get
the value already (IMHO there is no need for the overhead to first
lookup if the key is present and after that is checked, we call the get
function). It should be better to get the value and check if it differs
from null.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>
GitOrigin-RevId: 853179e8f28fcee2997b73b29df6efcdb6450c1d