[modbus] Workaround for nrjavaserial issues: do not disconnect serial (#2272)
Workaround for https://github.com/openhab/openhab-core/issues/1842.
Typically serial connections are disconnected and reconnected on IO
errors. This has turned out to be problematic with the recent
nrjavaserial updates brought by OH3 (locking behaviour of nrjavaserial
seems to be prone to deadlocks).

We workaround the issues by trying to keep the serial connection open
as long as possible.

Signed-off-by: Sami Salonen <ssalonen@gmail.com>
GitOrigin-RevId: b1b58436f2fb3aec8f54185c1437e84780f4264d