Upgrade Aries JAX-RS Whiteboard to 2.0.0 (#2316)
Upgrades the Aries JAX-RS Whiteboard from 1.0.9 to 2.0.0

The biggest change is that CXF is now an external dependency of the Aries JAX-RS Whiteboard bundle.
This also fixes the issue that you could not properly debug the CXF code because there was no source code for the embedded classes.
CXF 3.4.3 is now used instead of 3.2.14.

For release notes see:

* Changelog: https://github.com/apache/aries-jax-rs-whiteboard/blob/master/changelog.md
* Apache JIRA: https://issues.apache.org/jira/issues/?jql=project%20%3D%20ARIES%20AND%20fixVersion%20in%20(jax-rs-whiteboard-1.0.10%2C%20jax-rs-whiteboard-1.1.0%2C%20jax-rs-whiteboard-2.0.0)%20AND%20component%20%3D%20jax-rs-whiteboard

Signed-off-by: Wouter Born <github@maindrain.net>
GitOrigin-RevId: 0fe6b31f1977b998d05329028241b03eee795b24