Assume installed addons are compatible (#2975)
It was reported that after an upgrade from aprevious version installed addons from marketplaces are not shown in the add-ons pages. This is a result of the new compatibility check which excludes incompatible add-ons by default. Already installed add-ons must be excluded from this check otherwise they can't be uninstalled.

Signed-off-by: Jan N. Klug <github@klug.nrw>
GitOrigin-RevId: ad3a3c1caf78d766d1422d4e5527278f1ad8cecb