return the precise scheduled execution time for timers (#3182)
When writing tests and mocking time, it can be important that
execution_time match _exactly_ with what was sent to reschedule().
When re-calculating based on now and the expected remaining delay,
this might be off by a few milliseconds. We have the exact time
that was requested already, so might as well use it.

Signed-off-by: Cody Cutrer <cody@cutrer.us>
GitOrigin-RevId: d96e5d5bc01d19d9d11f782b4d1335237daf10e0